Major typechecking cleanup: use scalar doubles instead of floats when possible float arrays stay float Change some variable types, call different routines, and/or add lots of explicit casts to eliminate warnings from -Wno-int-conversion -Wno-sign-conversion -Wconversion main integer change: use size_t/ssize_t for buffer length arguments, not int Blocksize is now in seconds, back computed from user config value adjusted for clean FFT block sizes with front end sample rate Seems to mostly work but not fully tested, use at your own risk. Recommend running with assertions enabled for a while.

+++ src/aprs.c
@@ -212,7 +212,7 @@ int main(int argc,char *argv[]){
     exit(EX_NOINPUT);
   }
   uint8_t packet[PKTSIZE];
-  int size;
+  ssize_t size;
 
   while((size = recv(Input_fd,packet,sizeof(packet),0)) > 0){
     struct rtp_header rtp_header;
@@ -359,7 +359,7 @@ int main(int argc,char *argv[]){
 char *parse_timestamp(char *data,int *days,int *hours, int *minutes, int *seconds){
   // process timestamp
   char *ncp = NULL;
-  int t = strtol(data,&ncp,10);
+  int t = (int)strtol(data,&ncp,10);
   switch(*ncp){
   case 'h':
     // Hours, minutes, seconds
